+++ js/editchamado.js
@@ -109,9 +109,14 @@ $(function(){
 
 	// Função para excluir registros
 	function Excluir(){
+		let decisao = confirm("Deseja mesmo exluir o chamado?");
+		if (decisao==true) {
 		tbClientes.splice(indice_selecionado, 1);
 		localStorage.setItem("tbClientes", JSON.stringify(tbClientes));
 		alert("Registro excluído.");
+		}else{
+			alert("O chamado não foi excluido");
+		}
 	}
 
 	// Função para pesquisar cliente
